// frontend/src/app/models/call.enum.ts
export enum CallStatus {
  RINGING = 'RINGING',
  ONGOING = 'ONGOING',
  REJECTED = 'REJECTED',
  ENDED = 'ENDED'
}

export enum CallType {
  PRIVATE = 'PRIVATE',
  GROUP = 'GROUP'
}

export enum SignalType {
  OFFER = 'OFFER',
  ANSWER = 'ANSWER',
  CANDIDATE = 'CANDIDATE',
  RINGING = 'RINGING',
  HANGUP = 'HANGUP'
}